combine the xds-translator and xds-server into one xds runner
* primarily to reduce memory and convergence time
